Remove unused fix-lint tox environment
