+++ testing/code/test_source.py
@@ -744,3 +744,19 @@ something
 '''"""
     result = getstatement(1, source)
     assert str(result) == "'''\n'''"
+
+
+def test_getstartingblock_multiline():
+    class A(object):
+        def __init__(self, *args):
+            frame = sys._getframe(1)
+            self.source = _pytest._code.Frame(frame).statement
+
+    # fmt: off
+    x = A('x',
+          'y'
+          ,
+          'z')
+    # fmt: on
+    values = [i for i in x.source.lines if i.strip()]
+    assert len(values) == 4
