+++ tox.ini
@@ -151,14 +151,6 @@ commands =
     rm -rf /tmp/doc-exec*
     make regen
 
-[testenv:fix-lint]
-skipsdist = True
-usedevelop = True
-deps =
-    autopep8
-commands =
-    autopep8 --in-place -r --max-line-length=120 --exclude=test_source_multiline_block.py _pytest testing setup.py pytest.py
-
 [testenv:jython]
 changedir = testing
 commands =
